// src/api/generated/model/shiftRevenueEntryResponseEntryCategory.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * MiniCRM API
 * API Documentation for MiniCRM Application
 * OpenAPI spec version: 1.0.0
 */

export type ShiftRevenueEntryResponseEntryCategory = typeof ShiftRevenueEntryResponseEntryCategory[keyof typeof ShiftRevenueEntryResponseEntryCategory];


export const ShiftRevenueEntryResponseEntryCategory = {
  REGULAR: 'REGULAR',
  FLAT_RATE: 'FLAT_RATE',
  WEEKLY: 'WEEKLY',
} as const;